+++ src/Engine/Conditions.h
@@ -3,34 +3,13 @@
 #include <array>
 
 #include "Engine/Objects/ItemEnums.h"
+#include "Engine/Objects/CharacterEnums.h"
 
 #include "Utility/IndexedArray.h"
-#include "Utility/Segment.h"
 
-struct Character;
+class Character;
 
-enum class Condition : uint32_t {
-    CONDITION_CURSED = 0,
-    CONDITION_WEAK = 1,
-    CONDITION_SLEEP = 2,
-    CONDITION_FEAR = 3,
-    CONDITION_DRUNK = 4,
-    CONDITION_INSANE = 5,
-    CONDITION_POISON_WEAK = 6,
-    CONDITION_DISEASE_WEAK = 7,
-    CONDITION_POISON_MEDIUM = 8,
-    CONDITION_DISEASE_MEDIUM = 9,
-    CONDITION_POISON_SEVERE = 10,
-    CONDITION_DISEASE_SEVERE = 11,
-    CONDITION_PARALYZED = 12,
-    CONDITION_UNCONSCIOUS = 13,
-    CONDITION_DEAD = 14,
-    CONDITION_PETRIFIED = 15,
-    CONDITION_ERADICATED = 16,
-    CONDITION_ZOMBIE = 17,
-    CONDITION_GOOD = 18,
-};
-using enum Condition;
+// TODO(captainurist): Just hide all of this behind a sane interface.
 
 class EquipmentPair {
  public:
