+++ src/Engine/Localization.h
@@ -5,8 +5,8 @@
 
 #include "Engine/Objects/NPCEnums.h"
 #include "Engine/Objects/CharacterEnums.h"
-#include "Engine/Conditions.h"
 
+#include "Utility/Workaround/ToUnderlying.h"
 #include "Utility/IndexedArray.h"
