+++ src/GUI/CMakeLists.txt
@@ -20,8 +20,6 @@ set(GUI_HEADERS
 add_library(gui STATIC ${GUI_SOURCES} ${GUI_HEADERS})
 target_check_style(gui)
 
-target_link_libraries(gui
-                      gui_ui
-                      utility)
+target_link_libraries(gui gui_ui engine utility)
 
 add_subdirectory(UI)
