+++ src/GUI/UI/UIPopup.cpp
@@ -22,6 +22,7 @@
 #include "Engine/Spells/Spells.h"
 #include "Engine/Party.h"
 #include "Engine/Time.h"
+#include "Engine/Conditions.h"
 
 #include "Io/Mouse.h"
